# golang: Use a non-blocking send in broadcast