--- a/go/src/hashrocket/go-websocket-server/handler.go
+++ b/go/src/hashrocket/go-websocket-server/handler.go
@@ -4,6 +4,7 @@ import (
 	"io"
 	"log"
 	"sync"
+	"sync/atomic"
 
 	"golang.org/x/net/websocket"
 )
@@ -71,15 +72,23 @@ func (h *benchHandler) echo(ws *websocket.Conn, payload interface{}) error {
 func (h *benchHandler) broadcast(ws *websocket.Conn, payload interface{}) error {
 	result := BroadcastResult{Type: "broadcastResult", Payload: payload}
 
+	wg := sync.WaitGroup{}
+	var count int64
 	h.mutex.RLock()
 
-	for c, _ := range h.conns {
-		if err := websocket.JSON.Send(c, &WsMsg{Type: "broadcast", Payload: payload}); err == nil {
-			result.ListenerCount += 1
-		}
+	for c := range h.conns {
+		wg.Add(1)
+		go func(ws *websocket.Conn) {
+			if err := websocket.JSON.Send(ws, &WsMsg{Type: "broadcast", Payload: payload}); err == nil {
+				atomic.AddInt64(&count, 1)
+			}
+			wg.Done()
+		}(c)
 	}
 
 	h.mutex.RUnlock()
+	wg.Wait()
+	result.ListenerCount = int(count)
 
 	return websocket.JSON.Send(ws, &result)
 }