@jackc 2016-09-01
I think the performance regression is due to spinning up a new goroutine per broadcast per client. Goroutines are cheap, but not free.

There are a few idiomatic Go approaches. 

One would be to use the producer/consumer model. We could spin up N worker goroutines that read off a channel of outbound messages. 

Another would be to have one persistent goroutine per connection for the outbound send. This would be connected via channel.

We could also use aggressive write timeouts to alleviate the slow client problem, though that doesn't parallelize the sending of a broadcast.