@keegancsmith 2016-09-01
> I think the performance regression is due to spinning up a new goroutine per broadcast per client. Goroutines are cheap, but not free.

Yes I agree with you, this is almost certainly the reason. The cost of the goroutine may not be as bad if instead we did this test on a real network, not localhost. That would likely make this perform much better than the older implementation.

> One would be to use the producer/consumer model. We could spin up N worker goroutines that read off a channel of outbound messages.

This is probably the easiest thing to implement and would be performant. I'd be interested in seeing if it improves on localhost performance given we introduce some extra synchronisation (consuming the channel).

> Another would be to have one persistent goroutine per connection for the outbound send. This would be connected via channel.

There are some things to consider for that:
* Should it be a buffered or an unbuffered channel? (If buffered, how much?)
* How to communicate back if the send errored. (I assume we send a channel to reply on)
* How to do the cleanup nicely. Currently having a RW lock prevents closing a connection that we may be broadcasting to.