@jboelter 2016-09-01
You can avoid the repeated marshaling of the JSON data also:

```go
	msg := &WsMsg{Type: "broadcast", Payload: payload}
	data, err := json.Marshal(msg)

	h.mutex.RLock()
	for c := range h.conns {
		if err := websocket.Message.Send(c, data); err == nil {
```